added proper configuration selection

+++ mtp/usb/Interface.h
@@ -36,7 +36,7 @@ namespace mtp { namespace usb
 	public:
 		Interface(const libusb_interface_descriptor &interface): _interface(interface) { }
 
-		u8 GetIndex() const
+		int GetIndex() const
 		{ return _interface.bInterfaceNumber; }
 
 		int GetNameIndex() const
